Add error logging to workflows.
When a transition fails (example: workflow server cannot upload a new document to an FTP server), the corresponding task object is set to failed and a log message is added to the 'log' metaset.
This way, a failed workflow may be debugged without having direct access to the server's log files.