Remove GORM/Hibernate versioning from Session object.
Session is only updated when a user with an active session issues a command to the server, in which case the session is extended by the session timeout.
Requests to the server are short-lived (up to several seconds),  so It does not really matter which thread extends the session in the rare case that two concurrent requests try to do so.